Se crearon Exceptions y se aplicaron a TorneoServiceImpl

+++ src/main/java/com/torneos/futbol/service/TorneoServiceImpl.java
@@ -1,12 +1,16 @@
 package com.torneos.futbol.service;
 
+import com.torneos.futbol.exception.DataNullException;
+import com.torneos.futbol.exception.IdNotFoundException;
 import com.torneos.futbol.model.dto.TorneoDto;
 import com.torneos.futbol.model.entity.Torneo;
 import com.torneos.futbol.repository.TorneoRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -16,27 +20,52 @@ public class TorneoServiceImpl implements TorneoService {
     private final TorneoRepository torneoRepository;
     @Override
     public Torneo save(TorneoDto torneoDto) {
-        Torneo torneo = new Torneo();
-        torneo.setEquipos(null);
-        torneo.setNombre(torneoDto.getNombre());
-        torneo.setFechaInicio(torneoDto.getFechaInicio());
-        return torneoRepository.save(torneo);
+        if(torneoDto.getNombre() != null && torneoDto.getFechaInicio() != null) {
+            Torneo torneo = new Torneo();
+            torneo.setEquipos(null);
+            torneo.setNombre(torneoDto.getNombre());
+            torneo.setFechaInicio(torneoDto.getFechaInicio());
+            return torneoRepository.save(torneo);
+        } else if (torneoDto.getNombre() == null) {
+            throw new DataNullException("El nombre no puede ser null", HttpStatus.BAD_REQUEST.value());
+        }
+        else {
+            throw new DataNullException("La FechaInicio no puede ser null",HttpStatus.BAD_REQUEST.value());
+        }
     }
 
     @Override
-    public void delete(Torneo torneo) {
-        torneoRepository.delete(torneo);
+    public void delete(Integer id) {
+        Optional<Torneo> torneoOptional = torneoRepository.findById(id);
+
+        // Verificar si el Torneo está presente
+        if (torneoOptional.isPresent()) {
+            // Obtener el Torneo
+            Torneo torneo = torneoOptional.get();
+            // Eliminar el Torneo
+            torneoRepository.delete(torneo);
+        } else {
+            // Manejar el caso donde el Torneo no está presente
+            throw new IdNotFoundException("No se encontró el Torneo con el ID: " + id, HttpStatus.NOT_FOUND.value());
+        }
     }
 
     @Override
     public Torneo findById(Integer id) {
-        return torneoRepository.findById(id).orElse(null);
+        Optional<Torneo> torneoOptional = torneoRepository.findById(id);
+        if(torneoOptional.isPresent()){
+            return torneoOptional.get();
+        }
+        else {
+            throw new IdNotFoundException("No se encuentra ningun Torneo con el ID: " + id, HttpStatus.NOT_FOUND.value());
+        }
     }
 
     @Override
     public List<Torneo> findAll() {
         return torneoRepository.findAll();
     }
+
 
     @Override
     public Torneo update(Torneo torneo,Integer id) {
@@ -48,11 +77,11 @@ public class TorneoServiceImpl implements TorneoService {
                 return torneoRepository.save(torneo);
             } else {
                 // Manejar el caso en que el jugador no exista
-                throw new RuntimeException("El torneo con ID " + id + " no existe.");
+                throw new IdNotFoundException("El torneo con ID " + id + " no existe.",HttpStatus.NOT_FOUND.value());
             }
         }
         else {
-            throw new RuntimeException("El torneo con esta vacio");
+            throw new DataNullException("El Torneo no puede ser null",HttpStatus.BAD_REQUEST.value());
         }
     }
 }
